\name{hbutler-package}
\alias{hbutler-package}
\alias{hbutler}
\docType{package}
\title{
\packageTitle{hbutler}
}
\description{
\packageDescription{hbutler}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hbutler}
\packageIndices{hbutler}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{hbutler}

Maintainer: \packageMaintainer{hbutler}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
